Use LLVM_ENABLE_RUNTIMES to enable compiler-rt
Using LLVM_ENABLE_PROJECTS to enable compiler-rt is deprecated. The same
goes for libclc after the next pulldown. This commit changes the
configure script to use LLVM_ENABLE_PROJECTS to future-proof this.

Unlike the existing python code, it uses an array of runtime strings
which is joined with separators at the end, rather than manually
book-keeping a string and its separators. This should result in simpler
code, especially when we would have to check whether the string is empty
or not before prepending a separator. The runtimes variable could also
be switched to this mechanism in time.